requirements.txt
# Core data processing
pandas>=2.0.0
networkx>=3.0
numpy>=1.24.0

# Machine Learning (XGBoost + sklearn)
xgboost>=2.0.0
scikit-learn>=1.3.0

# Google Cloud (only needed for production BigQuery ingestion)
google-cloud-bigquery>=3.0.0

# Backend API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Frontend
streamlit>=1.28.0
pyvis>=0.3.2
plotly>=5.18.0
requests>=2.31.0

